+++ stream-chat-android-markdown-transformer/src/main/kotlin/io/getstream/chat/android/markdown/MarkdownTextTransformer.kt
@@ -3,7 +3,7 @@ package io.getstream.chat.android.markdown
 import android.content.Context
 import android.widget.TextView
 import com.getstream.sdk.chat.adapter.MessageListItem
-import io.getstream.chat.android.ui.transformer.AutoLinkableTextTransformer
+import com.getstream.sdk.chat.utils.Linkify
 import io.getstream.chat.android.ui.transformer.ChatMessageTextTransformer
 import io.noties.markwon.Markwon
 import io.noties.markwon.SoftBreakAddsNewLinePlugin
@@ -16,7 +16,7 @@ import io.noties.markwon.linkify.LinkifyPlugin
  * Markdown based implementation of [ChatMessageTextTransformer] that parses the message text as Markdown
  * and apply it to [TextView].
  */
-public class MarkdownTextTransformer(context: Context) : AutoLinkableTextTransformer {
+public class MarkdownTextTransformer(context: Context) : ChatMessageTextTransformer {
     private val markwon: Markwon = Markwon.builder(context)
         .usePlugin(CorePlugin.create())
         .usePlugin(LinkifyPlugin.create())
@@ -25,7 +25,8 @@ public class MarkdownTextTransformer(context: Context) : AutoLinkableTextTransfo
         .usePlugin(SoftBreakAddsNewLinePlugin.create())
         .build()
 
-    override fun transformer(textView: TextView, messageItem: MessageListItem.MessageItem) {
+    override fun transformAndApply(textView: TextView, messageItem: MessageListItem.MessageItem) {
         markwon.setMarkdown(textView, messageItem.message.text.fixItalicAtEnd())
+        Linkify.addLinks(textView)
     }
 }